LayoutTests/platform/wpe/fast/text/font-interstitial-invisible-width-while-loading-expected.txt

This test makes sure that the advances of the interstitial (invisible) font used during font loading is not very far from 'Times'. This is necessary for web compatibility.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS target.offsetWidth is within 5 of 222
PASS successfullyParsed is true

TEST COMPLETE
Hello
Hello
